+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## Version 26
+
+### v26.0.0
+
+- `DependsOnMethod` removed: use flat syntax with explicit method and a slash;
+- `options` property renamed to `ctx` in argument of:
+  - `Middleware::handler()`,
+  - `ResultHandler::handler()`,
+  - `handler` of `EndpointsFactory::build()` argument,
+  - `testMiddleware()`;
+- `EndpointsFactory::addOptions()` renamed to `addContext()`;
+
+```patch
+  const routing: Routing = {
+-   "/v1/users": new DependsOnMethod({
++   "/v1/users": {
+-     get: getUserEndpoint,
++     "get /": getUserEndpoint,
+-   }).nest({
+      create: makeUserEndpoint
+-   }),
++   },
+  };
+```
+
 ## Version 25
 
 ### v25.5.0